+++ app/lib/wagmi.ts
@@ -1,0 +1,15 @@
+import { createConfig, http } from 'wagmi';
+import { mainnet, sepolia } from 'wagmi/chains';
+import { injected } from 'wagmi/connectors';
+import { QueryClient } from '@tanstack/react-query';
+
+export const config = createConfig({
+  chains: [mainnet, sepolia],
+  connectors: [injected()],
+  transports: {
+    [mainnet.id]: http(),
+    [sepolia.id]: http(),
+  },
+});
+
+export const queryClient = new QueryClient();
